+++ package.json
@@ -25,7 +25,7 @@
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.2",
+    "react-scripts": "^4.0.3",
     "rehype-raw": "^5.0.0",
     "rehype-stringify": "^8.0.0",
     "remark-breaks": "^2.0.1",
