+++ src/containers/index.ts
@@ -1,9 +1,4 @@
 import AuthContainer from './AuthContainer';
 import BaseCoreContainer from './BaseCoreContainer';
-import LoginHeaderContainer from './LoginHeaderContainer';
 
-export {
-  AuthContainer,
-  BaseCoreContainer,
-  LoginHeaderContainer,
-};
+export { AuthContainer, BaseCoreContainer };
